✅ Task 6完了: PerformanceConfigurationIntegration.js分割 (64%削減、3コンポーネント)
🎯 Main Controller Pattern適用による分割:
- PerformanceConfigurationIntegration.js: 2,531語 → 918語 (64%削減)
- ConfigurationValidator.js: 1,354語 (検証・ルール管理)
- ConfigurationApplier.js: 1,587語 (設定適用・プロファイル管理)
- ConfigurationMonitor.js: 1,632語 (監視・統合・同期)

🔧 技術実装:
- Main Controller Pattern: 軽量オーケストレーター + サブコンポーネント
- 後方互換性: 既存公開API完全保持
- 依存注入: mainController参照による統制
- エラーハンドリング: ConfigErrorHandler統合

📊 成果:
- MCPツール互換性: 全ファイル2,500語以下達成
- アーキテクチャ: 単一責任原則適用
- 機能分離: バリデーション・適用・監視の分離
- API保持: updatePerformanceConfig等全メソッド維持

Requirements: 1.5, 8.1, 8.2
