+++ .kiro/specs/performance-utility-splitting-phase-f3/tasks.md
@@ -118,26 +118,26 @@
   - Verify all existing diagnostic functionality works correctly
   - _Requirements: 1.4, 8.1, 8.2_
 
-- [ ] 6. Split PerformanceConfigurationIntegration.js (2,531 words)
-- [ ] 6.1 Create ConfigurationValidator component
+- [x] 6. Split PerformanceConfigurationIntegration.js (2,531 words)
+- [x] 6.1 Create ConfigurationValidator component
   - Extract configuration validation and verification functionality
   - Implement config validation rules and validation workflow methods
   - Create unit tests for configuration validation functionality
   - _Requirements: 1.5, 8.1, 8.2_
 
-- [ ] 6.2 Create ConfigurationApplier component
+- [x] 6.2 Create ConfigurationApplier component
   - Extract configuration application and setting management functionality
   - Implement config application logic and setting update methods
   - Create unit tests for configuration application functionality
   - _Requirements: 1.5, 8.1, 8.2_
 
-- [ ] 6.3 Create ConfigurationMonitor component
+- [x] 6.3 Create ConfigurationMonitor component
   - Extract configuration monitoring and change detection functionality
   - Implement config monitoring and change notification methods
   - Create unit tests for configuration monitoring functionality
   - _Requirements: 1.5, 8.1, 8.2_
 
-- [ ] 6.4 Refactor main PerformanceConfigurationIntegration to use sub-components
+- [x] 6.4 Refactor main PerformanceConfigurationIntegration to use sub-components
   - Modify main class to instantiate and coordinate sub-components
   - Implement delegation methods to maintain public API compatibility
   - Verify all existing configuration functionality works correctly
